#1307 Moved CommonJobExecution to common package
